requirements.txt
cudatoolkit=10.1.243=h74a9793_0
gym=0.15.4=pypi_0
intel-openmp=2019.4=245
ipdb=0.12.3=py_0
ipykernel=5.1.2=py37h39e3cac_0
ipython=7.8.0=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
jupyter_client=5.3.4=py37_0
jupyter_core=4.6.0=py37_0
matplotlib=3.1.1=py37hc8f65d3_0
notebook=6.0.1=py37_0
numpy=1.16.5=py37h19fb1c0_0
numpy-base=1.16.5=py37hc3f5095_0
numpydoc=0.9.1=py_0
pandas=0.25.1=py37ha925a31_0
pip=19.2.3=py37_0
pysocks=1.7.1=py37_0
python=3.7.4=h5263a28_0
python-dateutil=2.8.0=py37_0
pytorch=1.3.1=py3.7_cuda101_cudnn7_0
scikit-learn=0.21.3=py37h6288b17_0
scipy=1.3.1=py37h29ff71c_0
spyder=3.3.6=py37_0
spyder-kernels=0.5.2=py37_0
torchcontrib=0.0.2=pypi_0